feat(notebooks)!: drop notebooks-collab-edit, keep notebook-edit only
`notebook-edit` (str_replace shape) covers every use case the
apply_patch tool was for, with a lower agent error rate in practice
(no line-prefix / EOL / fuzzy-matching footguns). Removing the second
tool keeps the model from having to pick between two ways of doing
the same thing, and drops ~1k lines of code we no longer need.

Removed:
  - src/tools/notebooks/collabEdit.ts    — apply_patch handler
  - src/tools/notebooks/patch.ts         — apply_patch parser
  - src/tools/notebooks/buildSteps.ts    — patch → steps
  - src/tools/notebooks/textRender.ts    — line-per-block PM rendering
  - schema/tool-definitions.json entry
  - TOOL_MAP entry in src/tools/index.ts

Retained:
  - schema.ts, rebase.ts, diffSteps.ts, saveLoop.ts — all used by
    notebook-edit
  - Schema + rebase unit tests, moved from notebooks-collab-edit.test.ts
    to a focused notebooks-internals.test.ts since they test shared
    primitives rather than any specific tool

This is a BREAKING change for anyone who had the previous unmerged
version of this branch — but since notebooks-collab-edit was never
in main, only branch consumers are affected.
